% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{borehole}
\alias{borehole}
\title{Create Borehole testing function
Source: Virtual Library of Simulation Experiments}
\usage{
borehole(xx)
}
\arguments{
\item{xx}{8-dim inputs}
}
\value{
water flow rates
}
\description{
Create Borehole testing function
Source: Virtual Library of Simulation Experiments
}
