% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{boxplot_cov}
\alias{boxplot_cov}
\title{Create rmse boxplots to compare covairance models}
\usage{
boxplot_cov(function_name, k, design_type, n, cov = "all", m = 20)
}
\arguments{
\item{function_name}{name of the testing function}

\item{k}{number of levels of the testing function}

\item{design_type}{the type of design: randomLHS, ortho_arrayLHS, maxiMin}

\item{n}{number of training and testing samples}

\item{cov}{covariance: all (all five possible cov types), gauss, matern5_2, matern3_2, exp, powexp}

\item{m}{number of repetitions}
}
\value{
the mean square error
}
\description{
Create rmse boxplots to compare covairance models
}
