% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{boxplot_design}
\alias{boxplot_design}
\title{Create rmse boxplots to compare design types}
\usage{
boxplot_design(
  function_name,
  k,
  design_type = "all",
  n,
  cov = "matern5_2",
  m = 20
)
}
\arguments{
\item{function_name}{name of the testing function}

\item{k}{number of levels of the testing function}

\item{design_type}{the type of design: all (all three design types), randomLHS, ortho_arrayLHS, maxiMin}

\item{n}{number of training and testing samples}

\item{cov}{covariance: gauss, matern5_2, matern3_2, exp, powexp}

\item{m}{number of repetitions}
}
\value{
the mean square error
}
\description{
Create rmse boxplots to compare design types
}
