% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{friedman}
\alias{friedman}
\title{Create FRIEDMAN FUNCTION
Source: https://www.sfu.ca/~ssurjano/fried.html}
\usage{
friedman(xx)
}
\arguments{
\item{xx}{5-dim inputs}
}
\value{
the Friedman function modeling result
}
\description{
Create FRIEDMAN FUNCTION
Source: https://www.sfu.ca/~ssurjano/fried.html
}
