% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{maxiMin}
\alias{maxiMin}
\title{Construct a maxiMin design}
\usage{
maxiMin(n, k)
}
\arguments{
\item{n}{the number of rows or samples}

\item{k}{the number of columns or parameters/variables}
}
\value{
A design matrix
}
\description{
Source: library(lhs)
Create a maxiMin design
}
