% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{ortho_arrayLHS}
\alias{ortho_arrayLHS}
\title{Construct an orthogonal array Latin hypercube}
\usage{
ortho_arrayLHS(n, k)
}
\arguments{
\item{n}{the number of rows or samples}

\item{k}{the number of columns or parameters/variables}
}
\value{
A design matrix
}
\description{
Source: library(lhs)
Create an orthogonal array Latin hypercube
}
