% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{otlcircuit}
\alias{otlcircuit}
\title{Create OTL Circuit testing function
Source: https://www.sfu.ca/~ssurjano/otlcircuit.html}
\usage{
otlcircuit(xx)
}
\arguments{
\item{xx}{6-dim inputs}
}
\value{
midpoint voltage
}
\description{
Create OTL Circuit testing function
Source: https://www.sfu.ca/~ssurjano/otlcircuit.html
}
