% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{piston}
\alias{piston}
\title{Create Piston testing function
Source: https://www.sfu.ca/~ssurjano/piston.html}
\usage{
piston(xx)
}
\arguments{
\item{xx}{7-dim inputs}
}
\value{
cycle time
}
\description{
Create Piston testing function
Source: https://www.sfu.ca/~ssurjano/piston.html
}
