% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{randomLHS}
\alias{randomLHS}
\title{Create a random Latin hypercube design}
\usage{
randomLHS(n, k)
}
\arguments{
\item{n}{the number of rows or samples}

\item{k}{the number of columns or parameters/variables}
}
\value{
A design matrix
}
\description{
Source: library(lhs)
randomLHS(4,3) returns a 4x3 matrix with each column constructed as follows:
A random permutation of (1,2,3,4) is generated, say (3,1,2,4) for each of K columns.
Then a uniform random number is picked from each indicated quartile one between 0 and .25,
then one between .25 and .5, finally one between .75 and 1.
}
