% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{rmse}
\alias{rmse}
\title{Testing the design models from scratch}
\usage{
rmse(function_name, k, design_type, n, cov = "matern5_2")
}
\arguments{
\item{function_name}{name of the testing function}

\item{k}{number of levels of the testing function}

\item{design_type}{the type of design: randomLHS, ortho_arrayLHS, maxiMin}

\item{n}{number of training/testing samples}

\item{cov}{the covariance type: gauss, matern5_2, matern3_2, exp, powexp}
}
\value{
the mean square error
}
\description{
Testing the design models from scratch
}
